Handle range value properly
Instead of relying on the fact that the provided value for a range query
is provided as float64, the util function is used to handle the cases
where the value is provided, i.e., as string.

This additionally improves the compatibility with Grafana since most
range query values will be sent as string.